refactoring the class: SqlParamBuilder
